+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 * Fixed session closing in `ydb.WithExecuteDataQueryOverQueryClient(true)` scenario
+
+## v3.110.1
 * Added the ability to send BulkRequest exceeding the GrpcMaxMessageSize
 
 ## v3.110.0